fix(talk): keep host error projection in the server runtime
